change: Event
I have defined the specification for the Event object. The Event object treats SDL events like a Single Table Inheritance (STI) model. It provides access to fields for all types of SDL events, returning either `nil` or valid values depending on the event type. In cases where SDL event fields involve memory management with pointers, the Event object automatically manages the pointers.

There are several ways SDL's event queue can only be accessed from the main thread. In this implementation, when relevant SDL functions are called, it monitors method calls from threads other than the main thread and returns exceptions.

Event.deq can be called from any thread. Use this method when implementing queue management in a multi-threaded context.

Event.get can only be called from the main thread and blocks if there are no events.

Event.wait waits for a new event if there are none in the queue. When called from a thread other than the main thread, it returns immediately without blocking.

Event.quit? can only be called from the main thread but allows for quick checking of shutdown notifications from the OS.

Event.update calls SDL_PumpEvents. It should be called regularly to retrieve events from the OS.

Operations for sending events to the SDL event queue are not implemented at the moment but will be implemented when needed.

Callback functions related to the event queue are not implemented. The reason is that callback functions do not work with Ruby's Global VM Lock (GVL) when called from other threads. In other words, we do not provide a way to peek into the event queue's contents. If you are interested in the contents of the event queue, retrieve all events on the Ruby side.