Implemented and Tested Mouse Module for SDL, with Important Note on Mouse State Retrieval
I have defined the specifications, implemented, and tested the Mouse module. The Mouse is an SDL internal singleton object that responds to methods regardless of which window it is in. In other words, it is not designed for applications composed of multiple windows. However, it is a very convenient feature for most applications, which typically have a single window.

As a crucial point, to know the latest state of the mouse, you need to call Event.update. This does not apply to methods with the global_ prefix; they directly query the OS.