Implemented Expected Functionality for RWOps in SDL
Implemented the expected functionality for RWOps in SDL, which is a crucial means for exchanging data with SDL or memory space from Ruby. There is potential for further extension in RWOps; however, this will have to wait for SDL3. At that time, RWOps is planned to be renamed as RW.